Moved map.vue to views
