+++ client/src/router.ts
@@ -2,7 +2,7 @@ import {createRouter, createWebHistory} from "vue-router";
 // Views
 import Home from "./views/Home.vue";
 import Results from "./views/Results.vue";
-import Map from "./components/Map.vue";
+import Map from "./views/Map.vue";
 // Authorization
 import Login from "./views/Login.vue";
 import Register from "./views/Register.vue";
